refactor: `ALIVE`, `DEAD`の定義位置を変更
