+++ app/src/utils/gol-patterns/random.ts
@@ -1,5 +1,4 @@
-import type { CellState } from "~/types/cell-state";
-import { ALIVE as A, DEAD as D } from "../gol-rule";
+import { ALIVE as A, DEAD as D, type CellState } from "~/types/cell-state";
 
 const randomize = (): CellState => (Math.random() >= 0.5 ? A : D);
