+++ app/tests/utils/gol-patterns/random.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it, mock, spyOn } from "bun:test";
+import { ALIVE, DEAD } from "~/src/types/cell-state";
 import { random } from "~/src/utils/gol-patterns/random";
-import { ALIVE, DEAD } from "~/src/utils/gol-rule";
 
 describe("random", () => {
   afterEach(() => {
